+++ core/src/executive/executive_tests.rs
@@ -1407,7 +1407,7 @@ fn test_storage_commission_privilege() {
         .unwrap();
     assert_eq!(
         state
-            .check_collateral_for_storage(
+            .check_collateral_for_storage_finally(
                 &privilege_control_address,
                 &U256::MAX,
                 &mut substate
@@ -1698,7 +1698,7 @@ fn test_storage_commission_privilege() {
     let mut substate = Substate::new();
     assert_eq!(
         state
-            .check_collateral_for_storage(
+            .check_collateral_for_storage_finally(
                 &privilege_control_address,
                 &U256::MAX,
                 &mut substate